refactor(renderer): R4 agents.js split + R3 unified create entry + R1 render-side cleanup
- agents.js (5138 lines) split into agents.js (1349, state host + list) / agents-detail.js (1742) / agents-create.js (531) / agents-picker.js (1556), loaded in dependency order per classic-script convention
- R3: createAgentViaIpc helper unified across 4 call sites (wizard x2, onboarding, cli-fallback)
- R1: 9 dead 'cli' checks removed (incl onboarding idempotency bug where the stale check never matched)
- test migration: shared multi-file loader helper; vm-style suites load ordered concatenation; text pins repointed to owning files; adoption-guard budget rebalanced per file
